Remove resolveFeatureFromFile and all use of it
I believe this used to run scenarios within Jest. Furthermore, I believe
this to be not only completely redundant, but also wrong.

Firstly, Cypress is the only interface for a user to consume this
library. No other test needs to pass than those in ./cypress.

Secondly, this happens to work because Jasmine aligns somewhat well with
Mocha, but this isn't guaranteed.

Lastly, but perhaps most importantly, this is in preperation of
upgrading to most recent version Gherkin, which only offers an async
API. Jasmine doesn't support deferred invocations of describe()
& it() like Mocha does.