refactor: upgrade to gherkin-9.0.0 and cucumber-head
This version of Gherkin outputs pickles for the JSON formatter to
consume, hence we update Cucumber as well. Cucumber's own update on
Gherkin is however unreleased and only resides in their master branch.

Since they do not commit the build to source code, we have to be
somewhat creative here. For now, one must clone the repository
recursively, install Cucumber's dependencies and build it.