Pass along all events to createTestsFromFeatures
This allows us to not parse the source again when generating JSON
reports. The observant reader will notice that we're removing the
dependency on "js-string-escape", which I believe is unnecessary when
using JSON.stringify as we do. The reason being that JS is a superset of
JSON in all relevant browsers [1].

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON#Browser_compatibility